Minor changes during PR review
